+++ CHANGELOG.md
@@ -16,3 +16,10 @@
 
 - Fixed few typos in README
 - Simplified the example and added README to /example folder
+
+## 1.0.4
+
+- `getStoreInstance` method now returns empty Store instance in case the value for given key was not found
+- Added `getNullableStoreInstance` method to StoreHolder
+- Added `removeOnDataChangedListener` method to Store
+- Updated README, example and tests